Rename file language-dropdown.vue to LanguageDropdown.vue
